rename teraNotebookError
